mise en place du Crud commentaire, finition du crud messages
Mise en anglais les termes utilisés.

+++ backend/controllers/message.js
@@ -1,25 +1,21 @@
 const db = require("../models");
-const Messages = db.messages;
-const User = db.user;
-const jwt = require('jsonwebtoken');
-const Op = db.Sequelize.Op;
+const Message = db.message;
 const fs = require('fs');
-const message = require("../models/message");
 
 // //creation d'un post//
 
 exports.createMessage = (req, res, next) => {
 
-  const messages = new Messages({
+  const message = new Message({
     title: req.body.title,
     text: req.body.text,
-    image: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
+    image: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`,
     userId: req.params.id,
     
   });
 
-  messages.save()
-  .then(() => res.status(201).json({ ...messages, messages: "message crée" }))
+  message.save()
+  .then(() => res.status(201).json({ ...message, message: "message crée" }))
   .catch(error => res.status(400).json({ error }));
 }
 
@@ -27,7 +23,7 @@ exports.createMessage = (req, res, next) => {
 
 exports.getOneMessages = (req, res, next) => {
 
-  Messages.findOne({ where: {id: req.params.id}, include: ["user"] })
+  Message.findOne({ where: {id: req.params.id}, include: ["user"] })
   .then(message => res.status(200).json(message))
   .catch( (error) =>  {res.status(404).json({
         error: error
@@ -35,11 +31,15 @@ exports.getOneMessages = (req, res, next) => {
 };
 
 exports.modifyMessage = (req, res, next) => {
+  //
   
+  if ( req.body.userId === req.body.id || req.body.role ==="admin"){
+
   if (req.file) {
-     Messages.findOne({where: {id: req.params.id}, include: ["user"]  })
-  .then(message => {
+     Message.findOne({where: {id: req.params.id}, include: ["user"]  }) 
  
+  .then(message => {
+
     const filename = message.image.split('/images/')[1]; 
      fs.unlinkSync(`images/${filename}`)
   }) 
@@ -54,32 +54,43 @@ const messageObject = req.file ?
 } : { ...req.body };
 console.log(messageObject); 
 //utilisation update afin de mettre a jours les données que l'utilisateur aura modifié. le where servant à dire où chercher.
-Messages.update({ ...messageObject, id:  req.params.id}, { where: { id: req.params.id }})
+Message.update({ ...messageObject, id:  req.params.id}, { where: { id: req.params.id }})
 //dans la response 200 ont renvoie la constante crée plus tôt avec les données modifié. dans le cas contraire une erreur 400 apparaitra.
 .then(() => res.status(200).json({ ...messageObject}))   
 .catch((error) => {res.status(400).json({ error })});
+  }
+  else{
+    res.send({
+      message: `Vous n'êtes pas autorisé à modifier ce message.`
+    });
+  }
 }
 
 exports.getAllMessages = (req, res, next) => {
-  Messages.findAll({ include: ["user"]} )
-    .then(message => res.status(200).json(message))
+  Message.findAll({ include: ["user"]} )
+    .then(messages => res.status(200).json(messages))
     .catch((error) => {res.status(400).json({ error })
   });
 }
 exports.deleteMessage = (req, res) => {
 
-//utilisation de deux condition if recupérant le role user et admin
-
-   //utilisation d'User.Destroy afin de supprimer l'utilisateur.
-   Messages.findOne({ where: {id: req.params.id} })
+//utilisation de deux condition if recupérant le role user et admin 
+if (req.body.id ===  req.body.userId || req.body.role === "admin"){
+   //utilisation de message.Destroy afin de supprimer le message
+   Message.findOne({ where: {id: req.params.id} })
    .then(message => {
      const filename = message.image.split('/images/')[1];
      fs.unlink(`images/${filename}`, () => {
-       Messages.destroy({where: {id: req.params.id} })
-         .then(() => res.status(200).json({ message: 'message supprimée !'}))
+       Message.destroy({where: {id: req.params.id} })
+         .then(() => res.status(200).json({ message: 'message supprimé !'}))
          .catch(error => res.status(400).json({ error }));
      });
    })
    .catch(error => res.status(500).json({ error }));
-
+  }
+  else{
+    res.send({
+      message: `Vous n'êtes pas autorisé à supprimer ce message.`
+    });
+  }
 }
